requirements.txt
awscli==1.32.39
boto3==1.34.39
contractions==0.1.73
emoji==2.10.1
jinja2==3.1.3
markdown==3.5.2
pandas==2.2.0
# `pattern` depends on gcc, mysql-devel and friends
pattern==3.6
praw==7.7.1
python-dotenv==1.0.1
# `nltk` should be installed before this for a smoother process
rake-nltk==1.0.6
psycopg2-binary==2.9.9
openai==1.12.0
backoff==2.2.1
tiktoken==0.6.0